+++ src/app/globals.css
@@ -1,4 +1,4 @@
-/* :root {
+:root {
   --max-width: 1100px;
   --border-radius: 12px;
   --font-mono: ui-monospace, Menlo, Monaco, 'Cascadia Mono', 'Segoe UI Mono',
@@ -104,4 +104,4 @@ a {
   html {
     color-scheme: dark;
   }
-} */
+}
