+++ src/pages/user/delete.tsx
